Race01 is reworked

+++ Race01/src/mx_strlen.c
@@ -1,7 +1,7 @@
-#include "../inc/minilibmx.h"
+#include "header.h"
 
-unsigned long mx_strlen(const char *s) {
-	unsigned long i = 0;
+int mx_strlen(const char *s) {
+	int i = 0;
 	for (; s[i] != '\0'; i++) {}
 	return i;
 }
